requirements.txt
absl-py==0.10.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.5.3
cycler==0.10.0
dask==0.18.2
decorator==4.3.0
face-alignment==1.1.1
facenet-pytorch==2.4.1
future==0.18.2
gast==0.3.3
google-auth==1.21.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.3.0
importlib-metadata==2.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.0.1
Markdown==3.2.2
matplotlib==2.2.2
mtcnn==0.1.0
networkx==2.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.3.0
pandas==0.23.4
Pillow==5.3.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.18
pygit==0.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
PyYAML==5.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
tifffile==2020.9.22
toolz==0.9.0
torch==1.2.0
torchvision==0.4.0
tqdm==4.24.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.2.0
